# api/src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Reception API",
    "description": "Implementation of AI Smart Contact Analysis in the NEAR blockchain"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": ""
    }
  ],
  "paths": {
    "/api/contract/analyze": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          }
        }
      }
    },
    "/api/contract/getUserAnalysis/{user_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/contract/getUsersAnalysis": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "UserRequest": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "string"
          },
          "contract": {
            "type": "string",
            "example": "string"
          },
          "api_key": {
            "type": "string",
            "example": "string"
          }
        },
        "xml": {
          "name": "UserRequest"
        }
      }
    }
  }
}